refactor main_settings_window in its own files
moved main_settings_window to a separate translation unit
